editoast: create object_type.rs in primitives folder
